# Conflicts:
#	ifs-web-service/ifs-competition-mgt-service/src/main/java/org/innovateuk/ifs/management/controller/CompetitionManagementAssessmentsAssessorProgressController.java
#	ifs-web-service/ifs-competition-mgt-service/src/main/java/org/innovateuk/ifs/management/model/AssessorAssessmentProgressModelPopulator.java
#	ifs-web-service/ifs-competition-mgt-service/src/main/java/org/innovateuk/ifs/management/viewmodel/AssessorAssessmentProgressViewModel.java
#	ifs-web-service/ifs-competition-mgt-service/src/main/resources/templates/competition/assessor-progress.html
#	ifs-web-service/ifs-competition-mgt-service/src/test/java/org/innovateuk/ifs/management/controller/CompetitionManagementAssessmentsAssessorProgressControllerTest.java

+++ ifs-resources/src/test/java/org/innovateuk/ifs/assessment/builder/AssessorCompetitionSummaryResourceBuilder.java
@@ -46,7 +46,7 @@ public class AssessorCompetitionSummaryResourceBuilder extends BaseBuilder<Asses
         return withArraySetFieldByReflection("competitionName", competitionNames);
     }
 
-    public AssessorCompetitionSummaryResourceBuilder withTotalApplications(Integer ...totalApplications) {
+    public AssessorCompetitionSummaryResourceBuilder withTotalApplications(Long ...totalApplications) {
         return withArraySetFieldByReflection("totalApplications", totalApplications);
     }
